requirements.txt
streamlit==1.28.1 
tensorflow-cpu==2.15.0 
pillow==10.0.1 
numpy==1.24.3 
pandas==2.0.3 
folium==0.14.0 
fpdf==1.7.2 
scikit-learn==1.3.0 
matplotlib==3.7.2 
